Extract shared types from legacy OnboardingView
- Promote OnboardingView.ViewState and its nested types (Intro, IntroType, IntroDialogType, StepInfo) to a standalone top-level OnboardingIntroViewState, so OnboardingIntroViewModel no longer
  depends on a view-layer type
- Remove the `extension OnboardingView { }` wrapper around DuckAIQuerySearchContent, making it a top-level struct; update the call site in RebrandedOnboardingView to drop the LegacyOnboardingView prefix
- Inline the daxGeometryEffectID string literal at its one rebranded call site in preparation for deleting the legacy constant

No behaviour change. Prepares for legacy OnboardingView deletion.
